Moving shared functions into a script helper
